python-service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
face-recognition==1.3.0
numpy==1.21.2
requests==2.26.0
imutils==0.5.4
python-dotenv==0.19.0
